commit changes: 1- changed control sizes of car ad cell 2- changed ad price to float 3- loaded car details data

+++ Bezaat/Bezaat/CarAds/Controller/CarAd.h
@@ -18,7 +18,7 @@
 @property (nonatomic) BOOL isFeatured;
 @property (strong, nonatomic) NSURL * thumbnailURL;
 @property (strong, nonatomic) NSString * title;
-@property (nonatomic) NSUInteger price;
+@property (nonatomic) float price;
 @property (strong, nonatomic) NSString * currencyString;
 @property (strong, nonatomic) NSDate * postedOnDate;
 @property (nonatomic) NSUInteger modelYear;
